Add safety check to graph remover

+++ catalogue_graph/src/graph_remover.py
@@ -2,9 +2,8 @@ import argparse
 import typing
 from datetime import datetime
 
-import polars as pl
-
 import config
+import polars as pl
 from transformers.create_transformer import EntityType, TransformerType
 from utils.aws import (
     df_from_s3_parquet,
@@ -18,6 +17,9 @@ IDS_LOG_SCHEMA: dict = {"timestamp": pl.Date(), "id": pl.Utf8}
 IDS_SNAPSHOT_FOLDER = "graph_remover/previous_ids_snapshot"
 DELETED_IDS_FOLDER = "graph_remover/deleted_ids"
 ADDED_IDS_FOLDER = "graph_remover/added_ids"
+
+# This is a safety mechanism. If two sets of IDs differ by more than 5%, an exception will be raised.
+ACCEPTABLE_DIFF_THRESHOLD = 0.05
 
 
 def get_previous_ids(
@@ -81,7 +83,7 @@ def log_ids(
 
 
 def delete_ids_from_neptune(
-    deleted_ids: set[str], entity_type: EntityType, is_local: bool
+    deleted_ids: set[str], entity_type: EntityType,  is_local: bool
 ) -> None:
     """Delete all nodes/edges with matching IDs from the Neptune cluster"""
     client = get_neptune_client(is_local)
@@ -92,7 +94,7 @@ def delete_ids_from_neptune(
 
 
 def handler(
-    transformer_type: TransformerType, entity_type: EntityType, is_local: bool = False
+    transformer_type: TransformerType, entity_type: EntityType, disable_safety_check: bool, is_local: bool = False
 ) -> None:
     try:
         # Retrieve a list of all ids which were loaded into the graph as part of the previous run
@@ -120,7 +122,10 @@ def handler(
             f"   Deleted ids: {len(deleted_ids)}\n",
             f"   Added ids: {len(added_ids)}",
         )
-
+        
+    if len(previous_ids) > 0 and len(deleted_ids) / len(previous_ids) > ACCEPTABLE_DIFF_THRESHOLD and not disable_safety_check:
+        raise ValueError(f"Attempted to remove {len(deleted_ids)} items (out of a total of {len(previous_ids)}), which is above the safety threshold.")
+    
     if len(deleted_ids) > 0:
         # Delete the corresponding items from the graph
         delete_ids_from_neptune(deleted_ids, entity_type, is_local)
@@ -138,7 +143,9 @@ def handler(
 def lambda_handler(event: dict, context: typing.Any) -> None:
     transformer_type = event["transformer_type"]
     entity_type = event["entity_type"]
-    handler(transformer_type, entity_type)
+    override_safety_check = event.get("override_safety_check", False)
+    
+    handler(transformer_type, entity_type, override_safety_check)
 
 
 def local_handler() -> None:
@@ -157,7 +164,13 @@ def local_handler() -> None:
         help="Which entity type to transform using the specified transformer (nodes or edges).",
         required=True,
     )
-
+    parser.add_argument(
+        "--disable-safety-check",
+        type=bool,
+        help="Whether to override a safety check which prevents node/edge removal if the percentage of removed entities is above a certain threshold.",
+        default=False
+    )
+    
     args = parser.parse_args()
 
     handler(**args.__dict__, is_local=True)
